Probably patch randomAttack
Now I haven't been able to see if this becomes out of sync because of rounding, but it shouldn't be. According to its page it's still not guaranteed these are in sync. The rounding is required to probably call upon the table. But either way the values should just be two types one is just an integer now but the same exact value without float.